Draft: CLI & Config
Tested some CLI, reworked the entry 'codecarto' imports.
Was getting a lot of circular imports, will have to figure out how this should be structured. 

A rework on how the config files and directories are created/set.
Was getting permission errors, but seems to be ok now. Need to test on system that does not currently have appdata codecarto folders/files