Draft: Database container, API responses, CSS Styles Update
API is working now

plotter, palette, and parser return something, even if something small for the moment.

TODO: palette needs a timed out catch to let the person know it's timed out.

database has also been included now. Nothing happening with it for the moment, but container is spinning up.